+++ nats-test/src/events/ticket-created-listener.ts
@@ -1,0 +1,13 @@
+import { Listener } from './base-listener';
+import { Message } from 'node-nats-streaming';
+
+export class TicketCreatedListener extends Listener {
+  subject = 'ticket:created';
+  queueGroupName = 'payments-service';
+
+  onMessage(data: any, msg: Message): void {
+    console.log('Event Data!', data);
+
+    msg.ack();
+  }
+}
